requirements.txt
absl-py==1.4.0
aio-pika==8.2.3
aiofiles==25.1.0
aiogram==2.25.1
aiohttp==3.8.6
aiohttp-retry==2.9.1
aiormq==6.4.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.9.1.post1
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.1.0
Babel==2.9.1
beautifulsoup4==4.14.3
bidict==0.23.1
boto3==1.42.22
botocore==1.42.22
bs4==0.0.2
CacheControl==0.12.14
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
confluent-kafka==2.13.0
cryptography==46.0.3
cycler==0.12.1
dask==2022.10.2
distro==1.9.0
dnspython==2.3.0
docopt==0.6.2
exceptiongroup==1.3.1
fbmessenger==6.0.0
filelock==3.20.2
fire==0.7.1
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.12.0
future==1.0.0
gast==0.4.0
google-auth==2.46.0
google-auth-oauthlib==1.0.0
google-genai==1.56.0
google-pasta==0.2.0
greenlet==3.3.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.2.4
humanfriendly==10.0
idna==3.11
jax==0.4.30
jaxlib==0.4.30
jmespath==1.0.1
joblib==1.5.3
jsonpickle==3.0.4
jsonschema==4.17.3
keras==2.12.0
kiwisolver==1.4.9
libclang==18.1.1
locket==1.0.0
magic-filter==1.0.12
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.5.3
mattermostwrapper==2.2
ml_dtypes==0.5.4
msgpack==1.1.2
multidict==5.2.0
networkx==2.6.3
numpy==1.23.5
oauthlib==3.3.1
opt_einsum==3.4.0
packaging==20.9
pamqp==3.2.1
partd==1.4.2
pillow==12.1.0
pluggy==1.6.0
portalocker==2.10.1
prompt-toolkit==3.0.28
propcache==0.4.1
protobuf==4.23.3
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydot==1.4.2
PyJWT==2.10.1
pykwalify==1.8.0
pymongo==4.3.3
pyparsing==3.3.1
pyrsistent==0.20.0
python-crfsuite==0.9.12
python-dateutil==2.8.2
python-engineio==4.13.0
python-socketio==5.16.0
pytz==2022.7.1
PyYAML==6.0.3
questionary==1.10.0
randomname==0.1.5
rasa==3.6.21
rasa-sdk==3.6.2
redis==4.6.0
regex==2022.10.31
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rocketchat-API==1.30.0
rsa==4.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.15
s3transfer==0.16.0
safetensors==0.4.5
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.1.3
scipy==1.10.1
sentry-sdk==1.14.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sklearn-crfsuite==0.3.6
skops==0.9.0
slack_sdk==3.39.0
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==1.4.54
structlog==23.3.0
structlog-sentry==2.1.0
tabulate==0.9.0
tarsafe==0.0.4
tenacity==9.1.2
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-text==2.12.0
termcolor==3.3.0
terminaltables==3.1.10
threadpoolctl==3.6.0
toolz==1.1.0
tqdm==4.67.1
twilio==8.2.2
typer-slim==0.21.1
typing-inspection==0.4.2
typing-utils==0.1.0
typing_extensions==4.15.0
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.6.2
uvloop==0.22.1
wcwidth==0.2.14
webexteamssdk==1.6.1
websockets==10.4
Werkzeug==3.1.4
wrapt==1.14.2
wsproto==1.3.2
yarl==1.22.0
